+++ src/plugins/Trace.ts
@@ -86,8 +86,6 @@ export class Trace {
                 propagatedSpanContext.transactionId : originalContext.awsRequestId;
 
             this.pluginContext.traceId = propagatedSpanContext.traceId;
-            this.pluginContext.spanId = propagatedSpanContext.spanId;
-
             this.tracer.transactionId = this.pluginContext.transactionId;
 
             this.rootSpan = this.tracer._startSpan(originalContext.functionName, {
@@ -108,18 +106,19 @@ export class Trace {
                 domainName: DomainNames.API,
                 className: ClassNames.LAMBDA,
             });
-
-            this.pluginContext.spanId = this.rootSpan.spanContext.spanId;
         }
 
+        this.pluginContext.spanId = this.rootSpan.spanContext.spanId;
+
         this.reporter = reporter;
-        this.startTimestamp = Date.now();
+        this.startTimestamp = this.pluginContext.invocationStartTimestamp;
+        this.rootSpan.startTime = this.pluginContext.invocationStartTimestamp;
 
         this.traceData = Utils.initMonitoringData(this.pluginContext,
             originalContext, MonitoringDataType.TRACE) as TraceData;
 
         this.traceData.id = this.pluginContext.traceId;
-        this.traceData.startTimestamp = Date.now();
+        this.traceData.startTimestamp = this.pluginContext.invocationStartTimestamp;
         this.traceData.rootSpanId = this.rootSpan ? this.rootSpan.spanContext.spanId : '';
 
         this.traceData.tags = {};
@@ -178,10 +177,11 @@ export class Trace {
 
         this.rootSpan.tags['aws.lambda.invocation.response'] = this.getResponse(response);
 
-        this.finishTimestamp = Date.now();
+        this.finishTimestamp = this.pluginContext.invocationFinishTimestamp;
         this.traceData.finishTimestamp = this.finishTimestamp;
         this.traceData.duration = this.finishTimestamp - this.startTimestamp;
         this.rootSpan.finish();
+        this.rootSpan.finishTime = this.pluginContext.invocationFinishTimestamp;
 
         const reportData = Utils.generateReport(this.traceData, this.apiKey);
         this.report(reportData);