+++ test/mocks/mocks.js
@@ -49,7 +49,9 @@ const createMockPluginContext = () => {
         applicationVersion: 'version',
         requestCount: 0,
         apiKey: 'apiKey',
-        maxMemory: 512
+        maxMemory: 512,
+        invocationStartTimestamp: Date.now(),
+        invocationFinishTimestamp: Date.now()
     };
 };