+++ test/reporter.test.js
@@ -1,31 +1,19 @@
 import Reporter from '../dist/Reporter.js';
-import {URL} from 'url';
 
-let httpRequestCalled = false;
 let httpsRequestCalled = false;
-
-let httpRequestOnCalled = false;
 let httpsRequestOnCalled = false;
-
-let httpRequestWriteCalled = false;
 let httpsRequestWriteCalled = false;
-
-let httpRequestEndCalled = false;
 let httpsRequestEndCalled = false;
 
-let httpSentData;
 let httpsSentData;
 
 jest.mock('http', () => ({
     request: (options, response) => {
-        httpRequestCalled = true;
         return {
-            on: jest.fn(() => httpRequestOnCalled = true),
+            on: jest.fn(() => true),
             write: jest.fn(data => {
-                httpRequestWriteCalled = true;
-                httpSentData = data;
             }),
-            end: jest.fn(() => httpRequestEndCalled = true)
+            end: jest.fn(() => true)
         };
     },
     Agent: () => {
@@ -52,10 +40,10 @@ jest.mock('https', () => ({
 
 describe('constructor', () => {
 
-    const reporter = new Reporter('apiKey');
+    const reporter = new Reporter({apiKey: 'apiKey'});
 
     it('should set api key', () => {
-        expect(reporter.apiKey).toEqual('apiKey');
+        expect(reporter.config.apiKey).toEqual('apiKey');
     });
 
     it('should set reports to empty array', () => {
@@ -68,10 +56,10 @@ describe('Reporter', () => {
 
     describe('constructor', () => {
 
-        const reporter = new Reporter('apiKey');
+        const reporter = new Reporter({apiKey: 'apiKey'});
 
         it('should set api key', () => {
-            expect(reporter.apiKey).toEqual('apiKey');
+            expect(reporter.config.apiKey).toEqual('apiKey');
         });
 
         it('should set reports to empty array', () => {
@@ -81,7 +69,7 @@ describe('Reporter', () => {
     });
 
     describe('addReports', () => {
-        const reporter = new Reporter('apiKey');
+        const reporter = new Reporter({apiKey: 'apiKey'});
         const mockReport = {data: 'data'};
         reporter.addReport(mockReport);
 
@@ -93,14 +81,18 @@ describe('Reporter', () => {
 
     describe('request', () => {
         describe('https', () => {
-            const reporter = new Reporter('apiKey');
+            const reporter = new Reporter({apiKey: 'apiKey'});
             const mockReport1 = {data: 'data1'};
             const mockReport2 = {data: 'data2'};
+
+            const reports = [];
+            reports.push(mockReport1);
+            reports.push(mockReport2);
 
             reporter.addReport(mockReport1);
             reporter.addReport(mockReport2);
             reporter.sendReports();
-
+            
             it('should make https request', () => {
                 expect(httpsRequestCalled).toEqual(true);
                 expect(httpsRequestOnCalled).toEqual(true);
@@ -109,38 +101,14 @@ describe('Reporter', () => {
             });
 
             it('should JSON.stringify reports on https.request', () => {
-                expect(httpsSentData).toEqual(JSON.stringify(reporter.reports));
-            });
-        });
-
-        describe('http', () => {
-            // noinspection JSAnnotator
-            const url = new URL('http://api.thundra.io/api');
-            const reporter = new Reporter('apiKey', url);
-            const mockReport1 = {data: 'data1'};
-            const mockReport2 = {data: 'data2'};
-
-            test('should make http request', () => {
-                reporter.addReport(mockReport1);
-                reporter.addReport(mockReport2);
-                reporter.sendReports();
-                expect(httpRequestCalled).toEqual(true);
-                expect(httpRequestOnCalled).toEqual(true);
-                expect(httpRequestWriteCalled).toEqual(true);
-                expect(httpRequestEndCalled).toEqual(true);
-
-            });
-
-            test('should JSON.stringify reports on https.request', () => {
-                expect(httpSentData).toEqual(JSON.stringify(reporter.reports));
+                expect(httpsSentData).toEqual(JSON.stringify(reports));
             });
         });
     });
-
-
+    
     describe('sendReports success', () => {
 
-        const reporter = new Reporter('apiKey');
+        const reporter = new Reporter({apiKey: 'apiKey'});
         const mockReport1 = {data: 'data1'};
         const mockReport2 = {data: 'data2'};
 
@@ -158,37 +126,11 @@ describe('Reporter', () => {
 
     });
 
-    describe('sendReports failure', () => {
-        
-        process.env.thundra_agent_lambda_debug_enable = 'true';
-        let consoleOutput;
-
-        const reporter = new Reporter('apiKey');
-        const mockReport1 = {data: 'data1'};
-        const mockReport2 = {data: 'data2'};
-
-        console['log'] = jest.fn(input => (consoleOutput = input));
-
-        reporter.request = jest.fn(async () => {
-            return {status: 400};
-        });
-
-        reporter.addReport(mockReport1);
-        reporter.addReport(mockReport2);
-        reporter.sendReports();
-
-        it('should log reports on failure', () => {
-            expect(reporter.request.mock.calls.length).toBe(1);
-            expect(consoleOutput).toEqual(JSON.stringify(reporter.reports));
-        });
-
-    });
-
     describe('addReports async', () => {
         let stdout = null;
         process.env.thundra_agent_lambda_report_cloudwatch_enable = true;
         process.stdout.write = jest.fn(input => stdout = input);
-        const reporter = new Reporter('apiKey');
+        const reporter = new Reporter({apiKey: 'apiKey'});
         const mockReport = {data: 'data'};
         reporter.addReport(mockReport);
         it('should not add report to reports array', () => {
@@ -197,7 +139,6 @@ describe('Reporter', () => {
         it('should write report to process.stdout', () => {
             expect(stdout).toBeTruthy();
         });
-
     });
 
 });